ischoolbus/requirements.txt
django==2.1.7
django-cors-headers
django-filter
django-mailer
django-ckeditor
django-mysql
djangorestframework
django-filer
fcm-django
python-slugify
python-dateutil
mysqlclient==1.3.13
coverage
pymysql
mysql-connector
gunicorn
markdown
flake8
yapf
django-mptt
easy_thumbnails
django-polymorphic
Pillow
djangorestframework-recursive
djangorestframework_simplejwt
django-daterangefilter
